Refactor patient record components
