+++ components/patientrecord/medical/PatientRecordDocumentListItem.vue
