+++ components/patientrecord/medical/PatientRecordMedicationCard.vue
