+++ components/patientrecord/medical/PatientRecordObservationCard.vue
