+++ components/patientrecord/medical/PatientRecordObservationRow.vue
