+++ components/patientrecord/medical/PatientRecordResultCard.vue
