+++ components/patientrecord/medical/PatientRecordResultRow.vue
