+++ components/workitem/WorkItemDetailCard.vue
