+++ pages/messages/_id/index.vue
@@ -128,9 +128,9 @@ import BaseInfoTooltip from "~/components/base/BaseInfoTooltip.vue";
 import BaseSkeleListItem from "~/components/base/BaseSkeleListItem.vue";
 import BaseSkeleText from "~/components/base/BaseSkeleText.vue";
 import MirthMessageListItem from "~/components/MirthMessageListItem.vue";
-import PatientRecordsListItem from "~/components/PatientRecordsListItem.vue";
+import PatientRecordsListItem from "~/components/patientrecord/PatientRecordsListItem.vue";
 import SendingFacilityLink from "~/components/SendingFacilityLink.vue";
-import WorkItemsListItem from "~/components/WorkItemsListItem.vue";
+import WorkItemsListItem from "~/components/workitem/WorkItemsListItem.vue";
 import useApi from "~/composables/useApi";
 import usePermissions from "~/composables/usePermissions";
 import useSensitive from "~/composables/useSensitive";
