+++ pages/patientrecords/_pid/documents/index.vue
@@ -32,7 +32,7 @@ import { DocumentSummarySchema, PatientRecordSchema } from "@ukkidney/ukrdc-axio
 import BaseCard from "~/components/base/BaseCard.vue";
 import BaseLoadingContainer from "~/components/base/BaseLoadingContainer.vue";
 import BasePaginator from "~/components/base/BasePaginator.vue";
-import PatientRecordDocumentListItem from "~/components/PatientRecordDocumentListItem.vue";
+import PatientRecordDocumentListItem from "~/components/patientrecord/medical/PatientRecordDocumentListItem.vue";
 import useDateRange from "~/composables/query/useDateRange";
 import usePagination from "~/composables/query/usePagination";
 import useApi from "~/composables/useApi";
