+++ pages/patientrecords/_pid/results.vue
@@ -125,8 +125,8 @@ import BaseModalConfirm from "~/components/base/BaseModalConfirm.vue";
 import BasePaginator from "~/components/base/BasePaginator.vue";
 import BaseSelectSearchable from "~/components/base/BaseSelectSearchable.vue";
 import BaseTable from "~/components/base/BaseTable.vue";
-import PatientRecordResultCard from "~/components/PatientRecordResultCard.vue";
-import PatientRecordResultRow from "~/components/PatientRecordResultRow.vue";
+import PatientRecordResultCard from "~/components/patientrecord/medical/PatientRecordResultCard.vue";
+import PatientRecordResultRow from "~/components/patientrecord/medical/PatientRecordResultRow.vue";
 import useDateRange from "~/composables/query/useDateRange";
 import usePagination from "~/composables/query/usePagination";
 import useQuery from "~/composables/query/useQuery";
